Reorganise error code
