+++ winterdrp/errors/exceptions.py
@@ -1,0 +1,14 @@
+class BaseProcessorError(Exception):
+    pass
+
+
+class ProcessorError(BaseProcessorError):
+    pass
+
+
+class NoncriticalProcessingError(BaseProcessorError):
+    pass
+
+
+class ImageNotFoundError(ProcessorError, FileNotFoundError):
+    pass
